requirements.txt
# Core framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.29
alembic==1.13.1

# Vector database
chromadb==0.5.0

# Embeddings & LLM clients
openai==1.30.0
anthropic==0.26.0
google-generativeai==0.5.4
tiktoken==0.7.0

# Utilities
httpx==0.27.0
numpy==1.26.4
tenacity==8.3.0
rich==13.7.1
typer==0.12.3

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
